Fix: Remove expenditure cancel handlers relying on old contracts
