+++ src/handlers/expenditures/helpers/index.ts
@@ -1,2 +1,1 @@
 export * from './getExpenditure';
-export * from './updateCancelledExpenditure';
