+++ src/handlers/expenditures/index.ts
@@ -12,7 +12,6 @@ export { default as handleExpenditurePayoutClaimed } from './expenditurePayoutCl
 export { default as handleStakeReclaimed } from './stakeReclaimed';
 export { default as handleExpenditureMadeStaged } from './expenditureMadeStaged';
 export { default as handleStagedPaymentReleased } from './stagedPaymentReleased';
-export { default as handleStakedExpenditureCancelled } from './stakedExpenditureCancelled';
 export { default as handleExpenditureStakerPunished } from './expenditureStakerPunished';
 export { default as handleStreamingPaymentCreated } from './streamingPaymentCreated';
 export { default as handleExpenditureMadeViaStake } from './expenditureMadeViaStake';
